#91 - Refactor: Extract class iterator from Tokens
