+++ src/Parser/Tokens.php
@@ -5,40 +5,6 @@
  * @license         http://www.opensource.org/licenses/bsd-license.php  BSD License *
  * @version         1.0                                                             */
 namespace Transphporm\Parser;
-
-class TokenFilterIterator implements \Iterator {
-    private $ignore;
-    private $tokens;
-
-    public function __construct(Tokens $tokens, array $ignore) {
-        $this->ignore = $ignore;
-        $this->tokens = $tokens;
-    }
-
-    public function current() {
-        return $this->tokens->current();
-    }
-
-    public function key() {
-        return $this->tokens->key();
-    }
-
-    public function valid() {
-        return $this->tokens->valid();
-    }
-
-    public function next() {
-        do {
-            $this->tokens->next();
-        }
-        while ($this->tokens->valid() && in_array($this->tokens->current()['type'], $this->ignore));
-    }
-
-    public function rewind() {
-        $this->tokens->rewind();
-        while ($this->tokens->valid() && in_array($this->tokens->current()['type'], $this->ignore)) $this->tokens->next();
-    }
-}
 class Tokens implements \Iterator, \Countable {
     private $tokens;
     private $iterator = 0;
@@ -129,16 +95,6 @@ class Tokens implements \Iterator, \Countable {
             array_shift($tokens);
         }
         return new Tokens($tokens);
-    }
-
-    public function getTokens() {
-    //    throw new \Exception();
-        $tokens = [];
-        //Loop through $this to account for $ignoreWhitespace
-        foreach ($this as $token) {
-            $tokens[] = $token;
-        }
-        return $tokens;
     }
 
     public function read($offset = 0) {
